# openapi.json
{"openapi":"3.1.0","info":{"title":"Картины на заказ","description":"Посетители создают заказ, художники его выполняют","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/Artists/{artist_id}":{"put":{"tags":["Artists"],"summary":"Update Artist","operationId":"update_artist_Artists__artist_id__put","parameters":[{"name":"artist_id","in":"path","required":true,"schema":{"type":"integer","title":"Artist Id"}},{"name":"new_style","in":"query","required":true,"schema":{"enum":["realism","impressionism","fauvism","modern","expressionism","cubism","futurism","abstractionism","dadaism","pop-art"],"type":"string","title":"New Style"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Artists"],"summary":"Get Artist By Id","operationId":"get_artist_by_id_Artists__artist_id__get","parameters":[{"name":"artist_id","in":"path","required":true,"schema":{"type":"integer","description":"ID of Artists to return","title":"Artist Id"},"description":"ID of Artists to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Artists":{"get":{"tags":["Artists"],"summary":"Find Artists By Style","operationId":"find_artists_by_style_Artists_get","parameters":[{"name":"style","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Style","description":"Style values that need to be considered for filter","enum":["realism","impressionism","fauvism","modern","expressionism","cubism","futurism","abstractionism","dadaism","pop-art"]},"description":"Style values that need to be considered for filter"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Artist"},"title":"Response Find Artists By Style Artists Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Vouchers":{"post":{"tags":["Vouchers"],"summary":"Place Vouchers","operationId":"place_vouchers_Vouchers_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Vouchers"],"summary":"Find Vouchers","operationId":"find_vouchers_Vouchers_get","parameters":[{"name":"style","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Style","description":"Style values that need to be considered for filter","enum":["realism","impressionism","fauvism","modern","expressionism","cubism","futurism","abstractionism","dadaism","pop-art"]},"description":"Style values that need to be considered for filter"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Status","description":"Status values that need to be considered for filter","enum":["placed","in work","ready"]},"description":"Status values that need to be considered for filter"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Voucher"},"title":"Response Find Vouchers Vouchers Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Vouchers/{voucher_id}":{"get":{"tags":["Vouchers"],"summary":"Get Vouchers By Id","operationId":"get_vouchers_by_id_Vouchers__voucher_id__get","parameters":[{"name":"voucher_id","in":"path","required":true,"schema":{"type":"integer","description":"ID of Vouchers to return","title":"Voucher Id"},"description":"ID of Vouchers to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Vouchers/{VouchersId}":{"put":{"tags":["Vouchers"],"summary":"Update Voucher","operationId":"update_voucher_Vouchers__VouchersId__put","parameters":[{"name":"voucher_id","in":"query","required":true,"schema":{"type":"integer","title":"Voucher Id"}},{"name":"new_status","in":"query","required":true,"schema":{"enum":["placed","in work","ready"],"type":"string","title":"New Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Vouchers"],"summary":"Delete Voucher","operationId":"delete_voucher_Vouchers__VouchersId__delete","parameters":[{"name":"VouchersId","in":"path","required":true,"schema":{"type":"integer","title":"Vouchersid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Accounts":{"post":{"tags":["Accounts"],"summary":"Place Accounts","operationId":"place_accounts_Accounts_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Accounts"],"summary":"Get Accounts By Name","operationId":"get_accounts_by_name_Accounts_get","parameters":[{"name":"first_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"The first name to search for.","title":"First Name"},"description":"The first name to search for."},{"name":"last_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"The last name to search for.","title":"Last Name"},"description":"The last name to search for."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"},"title":"Response Get Accounts By Name Accounts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Accounts"],"summary":"Update Account Name","operationId":"update_account_name_Accounts_put","parameters":[{"name":"new_first_name","in":"query","required":true,"schema":{"type":"string","description":"New first name for the account","title":"New First Name"},"description":"New first name for the account"},{"name":"new_last_name","in":"query","required":true,"schema":{"type":"string","description":"New last name for the account","title":"New Last Name"},"description":"New last name for the account"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Accounts"],"summary":"Delete Account","operationId":"delete_account_Accounts_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/Accounts/login":{"post":{"tags":["Accounts"],"summary":"Login","operationId":"login_Accounts_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Accounts/logout":{"get":{"tags":["Accounts"],"summary":"Logout","operationId":"logout_Accounts_logout_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/Visitors":{"put":{"tags":["Visitors"],"summary":"Update Visitor","operationId":"update_visitor_Visitors_put","parameters":[{"name":"visitor_id","in":"query","required":true,"schema":{"type":"integer","title":"Visitor Id"}},{"name":"new_residence","in":"query","required":true,"schema":{"type":"string","title":"New Residence"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Visitor"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/Visitors/{visitor_id}":{"get":{"tags":["Visitors"],"summary":"Get Visitor By Id","operationId":"get_visitor_by_id_Visitors__visitor_id__get","parameters":[{"name":"visitor_id","in":"path","required":true,"schema":{"type":"integer","description":"ID of Visitor to return","title":"Visitor Id"},"description":"ID of Visitor to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Visitor"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Account":{"properties":{"login":{"type":"string","title":"Login","example":"10"},"password":{"type":"string","title":"Password","example":"theAccounts"},"surName":{"type":"string","title":"Surname","example":"Green"},"firstName":{"type":"string","title":"Firstname","example":"John"},"patronymic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Patronymic","example":"James"},"email":{"type":"string","title":"Email","example":"john@email.com"},"type_role":{"type":"string","title":"Type Role","example":"artist"},"phone":{"type":"string","title":"Phone","example":"12345"},"sex":{"type":"string","enum":["f","m"],"title":"Sex","example":"m"},"date_of_birth":{"type":"string","title":"Date Of Birth","example":"2000-01-01"}},"type":"object","required":["login","password","surName","firstName","email","type_role","phone","sex","date_of_birth"],"title":"Account"},"Artist":{"properties":{"login":{"type":"string","title":"Login","example":"10"},"id":{"type":"integer","title":"Id","example":10},"style":{"type":"string","enum":["realism","impressionism","fauvism","modern","expressionism","cubism","futurism","abstractionism","dadaism","pop-art"],"title":"Style","example":"realism"}},"type":"object","required":["login","id","style"],"title":"Artist"},"Credentials":{"properties":{"login":{"type":"string","title":"Login","example":"login"},"password":{"type":"string","title":"Password","example":"password"}},"type":"object","required":["login","password"],"title":"Credentials"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Visitor":{"properties":{"login":{"type":"string","title":"Login","example":"10"},"id":{"type":"integer","title":"Id","example":10},"residence":{"type":"string","title":"Residence","example":"Yaroslavl, st/ Syrkova"}},"type":"object","required":["login","id","residence"],"title":"Visitor"},"Voucher":{"properties":{"id":{"type":"integer","title":"Id","example":10},"login":{"type":"string","title":"Login","example":"198login"},"amount_pictures":{"type":"integer","title":"Amount Pictures","example":3},"price":{"type":"integer","title":"Price","example":100},"description":{"type":"string","title":"Description","example":"3 pictures for 100$"},"status":{"type":"string","enum":["placed","in work","ready"],"title":"Status","example":"placed"},"style":{"type":"string","enum":["realism","impressionism","fauvism","modern","expressionism","cubism","futurism","abstractionism","dadaism","pop-art"],"title":"Style","example":"realism"}},"type":"object","required":["id","login","amount_pictures","price","description","status","style"],"title":"Voucher"}}}}